refactor(vue): register component-owned styles
