+++ packages/bamboo-vue/src/components/index.ts
@@ -1,16 +1,51 @@
-export { default as BoAccordion } from './bo-accordion/bo-accordion.vue';
-export { default as BoAccordionItem } from './bo-accordion/bo-accordion-item.vue';
-export { default as BoAlert } from './bo-alert/bo-alert.vue';
-export { default as BoAlertContainer } from './bo-alert/bo-alert-container.vue';
-export { default as BoAvatar } from './bo-avatar/bo-avatar.vue';
-export { default as BoBadge } from './bo-badge/bo-badge.vue';
-export { default as BoButton } from './bo-button/bo-button.vue';
+export { BoAccordion, BoAccordionItem } from './bo-accordion/index.js';
+export { BoAlert, BoAlertContainer } from './bo-alert/index.js';
+export { BoAlertDialog } from './bo-alert-dialog/index.js';
+export { BoAspectRatio } from './bo-aspect-ratio/index.js';
+export { BoAvatar } from './bo-avatar/index.js';
+export { BoBadge } from './bo-badge/index.js';
+export { BoBreadcrumb } from './bo-breadcrumb/index.js';
+export { BoButton } from './bo-button/index.js';
 export { BoButtonGroup } from './bo-button-group/index.js';
-export { default as BoIcon } from './bo-icon/bo-icon.vue';
-export { default as BoInput } from './bo-input/bo-input.vue';
-export { default as BoModal } from './bo-modal/bo-modal.vue';
-export { default as BoLoadingPulse } from './bo-loading-pulse/bo-loading-pulse.vue';
-export { default as BoLoadingSpinner } from './bo-loading-spinner/bo-loading-spinner.vue';
-export { default as BoText } from './bo-text/bo-text.vue';
-export { default as BoTextarea } from './bo-textarea/bo-textarea.vue';
+export { BoCard } from './bo-card/index.js';
+export { BoCheckbox } from './bo-checkbox/index.js';
+export { BoDialog } from './bo-dialog/index.js';
+export { BoEmpty } from './bo-empty/index.js';
+export { BoIcon } from './bo-icon/index.js';
+export { BoInput } from './bo-input/index.js';
+export { BoLabel } from './bo-label/index.js';
+export { BoModal } from './bo-modal/index.js';
+export { BoLoadingPulse } from './bo-loading-pulse/index.js';
+export { BoLoadingSpinner } from './bo-loading-spinner/index.js';
+export { BoMarker } from './bo-marker/index.js';
+export { BoPagination } from './bo-pagination/index.js';
+export { BoProgress } from './bo-progress/index.js';
+export { BoRadioGroup, BoRadioGroupItem } from './bo-radio-group/index.js';
+export { BoSeparator } from './bo-separator/index.js';
+export { BoSkeleton } from './bo-skeleton/index.js';
+export { BoTabs, BoTabsList, BoTabsTrigger, BoTabsContent } from './bo-tabs/index.js';
+export { BoText } from './bo-text/index.js';
+export { BoTextarea } from './bo-textarea/index.js';
+export { BoToast, BoToastViewport } from './bo-toast/index.js';
 export { BoTooltip } from './bo-tooltip/index.js';
+export { BoVisuallyHidden } from './bo-visually-hidden/index.js';
+
+export * from './bo-collapsible/index.js';
+export * from './bo-context-menu/index.js';
+export * from './bo-drawer/index.js';
+export * from './bo-hover-card/index.js';
+export * from './bo-input-otp/index.js';
+export * from './bo-menubar/index.js';
+export * from './bo-navigation-menu/index.js';
+export * from './bo-number-field/index.js';
+export * from './bo-pin-input/index.js';
+export * from './bo-popover/index.js';
+export * from './bo-resizable/index.js';
+export * from './bo-scroll-area/index.js';
+export * from './bo-select/index.js';
+export * from './bo-sheet/index.js';
+export * from './bo-sidebar/index.js';
+export * from './bo-switch/index.js';
+export * from './bo-tags-input/index.js';
+export * from './bo-toggle-group/index.js';
+export * from './bo-toggle/index.js';
