+++ packages/bamboo-vue/src/index.ts
@@ -3,25 +3,42 @@ import './lib.css';
 
 /** Component exports */
 
-export { default as BoAccordion } from './components/bo-accordion/bo-accordion.vue';
-export { default as BoAccordionItem } from './components/bo-accordion/bo-accordion-item.vue';
-export { default as BoAlert } from './components/bo-alert/bo-alert.vue';
-export { default as BoAlertContainer } from './components/bo-alert/bo-alert-container.vue';
-export { default as BoAvatar } from './components/bo-avatar/bo-avatar.vue';
-export { default as BoBadge } from './components/bo-badge/bo-badge.vue';
-export { default as BoButton } from './components/bo-button/bo-button.vue';
+export { BoAccordion, BoAccordionItem } from './components/bo-accordion/index.js';
+export { BoAlert, BoAlertContainer } from './components/bo-alert/index.js';
+export { BoAspectRatio } from './components/bo-aspect-ratio/index.js';
+export { BoAlertDialog } from './components/bo-alert-dialog/index.js';
+export { BoAvatar } from './components/bo-avatar/index.js';
+export { BoBadge } from './components/bo-badge/index.js';
+export { BoBreadcrumb } from './components/bo-breadcrumb/index.js';
+export { BoButton } from './components/bo-button/index.js';
 export { BoButtonGroup } from './components/bo-button-group/index.js';
-export { default as BoIcon } from './components/bo-icon/bo-icon.vue';
-export { default as BoInput } from './components/bo-input/bo-input.vue';
-export { default as BoLoadingPulse } from './components/bo-loading-pulse/bo-loading-pulse.vue';
-export { default as BoLoadingSpinner } from './components/bo-loading-spinner/bo-loading-spinner.vue';
-export { default as BoModal } from './components/bo-modal/bo-modal.vue';
-export { default as BoText } from './components/bo-text/bo-text.vue';
-export { default as BoTextarea } from './components/bo-textarea/bo-textarea.vue';
+export { BoCard } from './components/bo-card/index.js';
+export { BoCheckbox } from './components/bo-checkbox/index.js';
+export { BoDialog } from './components/bo-dialog/index.js';
+export { BoEmpty } from './components/bo-empty/index.js';
+export { BoIcon } from './components/bo-icon/index.js';
+export { BoInput } from './components/bo-input/index.js';
+export { BoLabel } from './components/bo-label/index.js';
+export { BoLoadingPulse } from './components/bo-loading-pulse/index.js';
+export { BoLoadingSpinner } from './components/bo-loading-spinner/index.js';
+export { BoMarker } from './components/bo-marker/index.js';
+export { BoModal } from './components/bo-modal/index.js';
+export { BoPagination } from './components/bo-pagination/index.js';
+export { BoProgress } from './components/bo-progress/index.js';
+export { BoRadioGroup, BoRadioGroupItem } from './components/bo-radio-group/index.js';
+export { BoSeparator } from './components/bo-separator/index.js';
+export { BoSkeleton } from './components/bo-skeleton/index.js';
+export { BoTabs, BoTabsList, BoTabsTrigger, BoTabsContent } from './components/bo-tabs/index.js';
+export { BoText } from './components/bo-text/index.js';
+export { BoTextarea } from './components/bo-textarea/index.js';
+export { BoToast, BoToastViewport } from './components/bo-toast/index.js';
 export { BoTooltip } from './components/bo-tooltip/index.js';
+export { BoVisuallyHidden } from './components/bo-visually-hidden/index.js';
 
 /** Plugin exports */
 
 export { BambooVue } from './plugin.js';
 export { useAlert } from './composables/use-alert.js';
 export type { AlertMessage, AlertOptions, AlertPosition } from './composables/use-alert.js';
+
+export * from './components/index.js';
